+++ doc/CHANGELOG.md
@@ -21,10 +21,12 @@ If you want to contribute to translating work, we manage the translation of this
 ### Feature Changes
 
 * Now whenever you're looking at a block and entity, the mod will play a piano sound cue to indicate relative location between you and target. see also: [description of this feature](/doc/FEATURES.md#relative-position-sound-cue) [#191](https://github.com/khanshoaib3/minecraft-access/issues/191)
+* Now the mod will only speak item count when held item type isn't changed (when picking up or spending items), with a new config "Report Held Items Count When Changed", see the [configuration](/doc/CONFIG.md#other-configurations).
 
 ### Bug Fixes
 
 * Fix the bug that keybindings can't be rebound to mouse 4,5 keys (mouse side keys) [#190](https://github.com/khanshoaib3/minecraft-access/issues/190)
+* Fix the bug that a dead entity can be locked before it disappears. [#208](https://github.com/khanshoaib3/minecraft-access/issues/208)
 
 ### Dependencies Changes